+++ src/components/Select/Select.tsx
@@ -19,7 +19,8 @@ export const labelPositionsArray = ["default", "inline"];
 export type SelectTypes = typeof selectTypesArray[number];
 export type LabelPositions = typeof labelPositionsArray[number];
 
-export interface SelectProps {
+export interface SelectProps
+  extends React.SelectHTMLAttributes<HTMLSelectElement> {
   /** A class name for the `div` parent element. */
   className?: string;
   /** The initial value of an uncontrolled component */
@@ -52,6 +53,9 @@ export interface SelectProps {
   onChange?: (event: React.FormEvent) => void;
   /** Placeholder text in the select element. */
   placeholder?: string;
+  /** Allows the '(required)' text to be changed for language purposes
+   * Note: Parenthesis will be added automatically by the component */
+  requiredLabelText?: string;
   /** The variant to display. */
   selectType?: SelectTypes;
   /** Offers the ability to hide the helper/invalid text. */
@@ -81,6 +85,7 @@ export const Select: ChakraComponent<
   forwardRef<HTMLSelectElement, React.PropsWithChildren<SelectProps>>(
     (props: React.PropsWithChildren<SelectProps>, ref?) => {
       const {
+        autoComplete,
         children,
         className,
         defaultValue,
@@ -99,6 +104,7 @@ export const Select: ChakraComponent<
         showHelperInvalidText = true,
         showLabel = true,
         showRequiredLabel = true,
+        requiredLabelText,
         value = "",
         ...rest
       } = props;
@@ -182,22 +188,22 @@ export const Select: ChakraComponent<
                   id={`${id}-label`}
                   isInlined
                   isRequired={showRequiredLabel && isRequired}
+                  requiredLabelText={requiredLabelText}
                 >
                   {labelText}
                 </Label>
               </Box>
             )}
             <ChakraSelect
+              autoComplete={autoComplete}
               id={id}
-              variant="outline"
               isRequired={isRequired}
               isDisabled={isDisabled}
               isInvalid={isInvalid}
               name={name}
               placeholder={placeholder}
               ref={ref}
-              {...controlledOrUncontrolledProps}
-              {...ariaAttributes}
+              variant="outline"
               icon={
                 <Icon
                   color={arrowColor}
@@ -206,6 +212,8 @@ export const Select: ChakraComponent<
                   size="medium"
                 />
               }
+              {...controlledOrUncontrolledProps}
+              {...ariaAttributes}
               __css={styles.select}
             >
               {children}