+++ src/components/TagSet/tagSetChangelogData.ts
@@ -10,6 +10,24 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: [
+      "Fixes an overflow bug in the `filter` variant when `isDismissible` is false.",
+    ],
+  },
+  {
+    date: "2025-03-20",
+    version: "3.5.5",
+    type: "Update",
+    affects: ["Documentation", "Accessibility"],
+    notes: [
+      "Updated the `'filter'` variant to remove button wrapper on each tag when `isDismissible` is false and no `onClick` is passed.",
+    ],
+  },
+  {
     date: "2024-04-25",
     version: "3.1.1",
     type: "Update",